TaskManager_back/pyproject.toml
[project]
name = "taskmanager_back"
version = "0.1.0"
description = "Task Manager BackEnd"
authors = [
    {name = "slavaklevleev",email = "slava.klevleev@gmail.com"}
]
package-mode = false 
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
